Renamed patched.rb to the original asciidoctor-pdf-mathjax.rb
