+++ lib/asciidoctor-pdf-mathjax.rb
@@ -5,14 +5,21 @@ require 'rexml/document'
 require 'ttfunk'
 require 'asciimath'
 
+require 'zlib'
+
 POINTS_PER_EX = 6
-MATHJAX_DEFAULT_COLOR_STRING = "currentColor"
-MATHJAX_DEFAULT_FONT_FAMILY = "TeX"
+MATHJAX_DEFAULT_COLOR_STRING = 'currentColor'
+MATHJAX_DEFAULT_FONT_FAMILY = 'TeX'
 
 FALLBACK_FONT_SIZE = 12
 FALLBACK_FONT_STYLE = 'normal'
 FALLBACK_FONT_FAMILY = 'Arial'
 FALLBACK_FONT_COLOR = '#000000'
+
+ATTRIBUTE_FONT = 'math-font'
+ATTRIBUTE_CACHE_DIR = 'math-cache-dir'
+PREFIX_STEM = 'stem-'
+PREFIX_WIDTH = 'width-' # width cache files
 
 class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
   register_for 'pdf'
@@ -22,19 +29,193 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
     attr_reader :tempfiles
   end
 
+  ##### patch start
+  @@cached_svg_width = {}
+  @@cached_svg_output = {}
+  @@cache_dir_init_done = false
+
+  def get_hash_key(latex_content, math_font, is_inline)
+    b = (is_inline ? 'true' : 'false')
+    data = latex_content + math_font + b
+    Zlib.adler32(data).to_s(16).freeze
+  end
+
+  def get_adjusted_svg_from_node(node, latex_content, is_inline, math_font)
+    svg_output, error = stem_to_svg(latex_content, is_inline, math_font)
+
+    if svg_output == ''
+      s = "No svg produced when adjusting LaTeX:\n" + latex_content
+      logger.error(s)
+      error = s
+    end
+
+    return nil, error unless error.nil?
+
+    if !is_inline
+      svg_output = adjust_svg_color(svg_output, @font_color)
+      svg_default_font_size = FALLBACK_FONT_SIZE
+
+      svg_doc = REXML::Document.new(svg_output)
+      svg_width = (svg_doc.root.attributes['width'].to_f * POINTS_PER_EX) || raise('No width found in SVG')
+
+      scaling_factor = @font_size.to_f / svg_default_font_size
+      svg_width *= scaling_factor
+
+      [svg_output: svg_output, svg_width: svg_width]
+    else
+      theme = (load_theme node.document)
+      svg_output, svg_width = adjust_svg_to_match_text(svg_output, node, theme)
+      [svg_output: svg_output, svg_width: svg_width]
+    end
+  end
+
+  def get_adjusted_svg_and_set_cached_width(node, latex_content, is_inline, math_font, hash_key)
+    adjusted_svg, error = get_adjusted_svg_from_node(node, latex_content, is_inline, math_font)
+
+    return nil, error unless error.nil?
+
+    @@cached_svg_width[hash_key] = adjusted_svg[:svg_width]
+
+    adjusted_svg
+  end
+
+  def get_cached_svg_file_path(cache_dir, hash_key)
+    File.join(cache_dir, PREFIX_STEM + hash_key + '.svg')
+  end
+
+  def get_cached_svg_width_path(cache_dir, hash_key)
+    File.join(cache_dir, PREFIX_WIDTH + hash_key)
+  end
+
+  def get_cached_svg_width(cache_dir, hash_key)
+    unless @@cached_svg_width[hash_key] # not cached; read from file
+      file_name = get_cached_svg_width_path(cache_dir, hash_key)
+      svg_width = File.read(file_name).to_f
+      L('Returning cached file width: ' + svg_width.to_s + ' for hash_key: ' + hash_key)
+      @@cached_svg_width[hash_key] = svg_width.freeze
+    end
+
+    svg_width = @@cached_svg_width[hash_key]
+    L('Returning cached ram width: ' + svg_width.to_s + ' for hash_key: ' + hash_key)
+    svg_width
+  end
+
+  def get_cached_svg_output(cache_dir, hash_key)
+    unless @@cached_svg_output[hash_key] # read from file
+      file_name = get_cached_svg_file_path(cache_dir, hash_key)
+      svg_content = File.read(file_name)
+      @@cached_svg_output[hash_key] = svg_content.freeze
+    end
+
+    @@cached_svg_output[hash_key]
+  end
+
+  def get_math_font(node)
+    node.document.attributes[ATTRIBUTE_FONT] || MATHJAX_DEFAULT_FONT_FAMILY
+  end
+
+  def L(debug_text)
+    logger.debug('PATCH: ' + debug_text)
+  end
+
+  # return {file_path, svg_width, temp_handle, inline_nil_latex, inline_nil_svg}
+  def get_svg_info(node, is_inline)
+    if is_inline
+      node_arg1 = node.text
+      node_arg2 = node.type
+    else
+      node_arg1 = node.content
+      node_arg2 = node.style.to_sym
+    end
+    latex_content = extract_latex_content(node_arg1, node_arg2)
+
+    return { svg_output: '', latex_content: '' } if latex_content.nil?
+
+    empty_reply = { svg_output: '', latex_content: latex_content }
+
+    math_font = get_math_font(node).freeze
+    cache_dir = (node.document.attributes[ATTRIBUTE_CACHE_DIR] || nil).freeze
+
+    hash_key = get_hash_key(latex_content, math_font, is_inline)
+
+    unless cache_dir.nil? # caching enabled
+      unless @@cache_dir_init_done # ensure directory exists
+        puts('Cache directory set to: ' + cache_dir)
+        L('INIT cache dir: ' + cache_dir)
+        @@cache_dir_init_done = true
+        FileUtils.mkdir_p(cache_dir) unless Dir.exist?(cache_dir)
+      end
+
+      cached_svg_file_path = get_cached_svg_file_path(cache_dir, hash_key)
+
+      unless @@cached_svg_output[hash_key].nil? # read from memory
+        svg_width = get_cached_svg_width(cache_dir, hash_key)
+        svg_output = get_cached_svg_output(cache_dir, hash_key)
+        L('returning RAM cached data for hash_key: ' + hash_key)
+        return { file_path: cached_svg_file_path, svg_output: svg_output, svg_width: svg_width,
+                 latex_content: latex_content }
+      end
+
+      if File.exist?(cached_svg_file_path)
+        svg_width = get_cached_svg_width(cache_dir, hash_key)
+        svg_output = get_cached_svg_output(cache_dir, hash_key)
+        L('returning DISK cached data for hash_key: ' + hash_key)
+        return { file_path: cached_svg_file_path, svg_output: svg_output, svg_width: svg_width,
+                 latex_content: latex_content }
+      end
+    end
+
+    # caching disabled, or file doesn't exist in cache yet, so create
+    L('GENERATING data for hash_key:' + hash_key + ' -- Latex content: ' + latex_content)
+
+    adjusted_svg, error = get_adjusted_svg_and_set_cached_width(node, latex_content, is_inline, math_font, hash_key)
+    return empty_reply, error if error
+
+    svg_output = adjusted_svg[:svg_output]
+    svg_width = adjusted_svg[:svg_width]
+
+    unless cache_dir.nil? # cache to disk the generated svg and width data
+      L('Writing svg content and width to RAM')
+      @@cached_svg_output[hash_key] = svg_output
+      @@cached_svg_width[hash_key] = svg_width
+
+      L('Writing svg content and width to DISK')
+      File.write(cached_svg_file_path, svg_output)
+      cached_svg_width_file_path = get_cached_svg_width_path(cache_dir, hash_key)
+      File.write(cached_svg_width_file_path, svg_width)
+
+      L('returning NEWLY CACHED path, width, and content for hash_key: ' + hash_key)
+      return { file_path: cached_svg_file_path, svg_output: svg_output, svg_width: svg_width,
+               latex_content: latex_content }
+    end
+
+    # no caching, use original Tempfile method.
+    L('no caching.  writing to tempfile')
+    file_handle = Tempfile.new(['stem', '.svg'])
+    self.class.tempfiles << file_handle
+    file_handle.write(svg_output)
+    file_handle.close
+    # no unlinking here.  unlink after the temp file has been used.
+
+    L('returning uncached temp file path, and svg width')
+    { file_path: file_handle.file_path, svg_output: svg_output, svg_width: svg_width,
+      temp_handle: file_handle }
+  end
+  # end of patch
+
   def convert_stem(node)
     arrange_block node do |_|
       add_dest_for_block node if node.id
 
-      latex_content = extract_latex_content(node.content, node.style.to_sym)
-      math_font = node.document.attributes['math-font'] || MATHJAX_DEFAULT_FONT_FAMILY
-
-      svg_output, error = stem_to_svg(latex_content, false, math_font)
+      svg_info, error = get_svg_info(node, false)
+      svg_output = svg_info[:svg_output]
+      latex_content = svg_info[:latex_content]
 
       # noinspection RubyResolve
       code_padding = @theme.code_padding
-      if svg_output.nil? || svg_output.empty?
-        logger.warn "Failed to convert STEM to SVG: #{error} (Fallback to code block)"
+      if svg_output == ''
+        # logger.warn "Failed to convert STEM to SVG: #{error} (Fallback to code block)"
+        logger.warn 'Failed to convert STEM to SVG. (Fallback to code block)'
         pad_box code_padding, node do
           theme_font :code do
             typeset_formatted_text [{ text: (guard_indentation latex_content), color: @font_color }],
@@ -43,32 +224,27 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
           end
         end
       else
-        svg_output = adjust_svg_color(svg_output, @font_color)
-        svg_default_font_size = FALLBACK_FONT_SIZE
+        # removed width and temp file code
 
-        svg_doc = REXML::Document.new(svg_output)
-        svg_width = svg_doc.root.attributes['width'].to_f * POINTS_PER_EX || raise("No width found in SVG")
-
-        scaling_factor = @font_size.to_f / svg_default_font_size
-        svg_width = svg_width * scaling_factor
-
-        svg_file = Tempfile.new(%w[stem .svg])
         begin
-          svg_file.write(svg_output)
-          svg_file.close
+          # removed temp file writing and close
+
+          svg_width = svg_info[:svg_width]
+          file_path = svg_info[:file_path]
+          temp_handle = svg_info[:temp_handle]
+          L('Prawn embed path: ' + file_path)
+          L('Prawn embed width:' + svg_width.to_s)
 
           pad_box code_padding, node do
-            begin
-              image_obj = image svg_file.path, position: :center, width: svg_width, height: nil
-              logger.debug "Successfully embedded stem block (as latex) #{latex_content} as SVG image" if image_obj
-            rescue Prawn::Errors::UnsupportedImageType => e
-              logger.warn "Unsupported image type error: #{e.message}"
-            rescue StandardError => e
-              logger.warn "Failed embedding SVG: #{e.message}"
-            end
+            image_obj = image file_path, position: :center, width: svg_width, height: nil
+            logger.debug "Successfully embedded stem block (as latex) #{latex_content} as SVG image" if image_obj
+          rescue Prawn::Errors::UnsupportedImageType => e
+            logger.warn "Unsupported image type error: #{e.message}"
+          rescue StandardError => e
+            logger.warn "Failed embedding SVG: #{e.message}"
           end
         ensure
-          svg_file.unlink
+          temp_handle.unlink if temp_handle
         end
       end
     end
@@ -76,28 +252,31 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
   end
 
   def convert_inline_quoted(node)
-    latex_content = extract_latex_content(node.text, node.type)
-    return super if latex_content.nil?
+    svg_info, error = get_svg_info(node, true)
+    svg_output = svg_info[:svg_output]
+    latex_content = svg_info[:latex_content]
+    return super if latex_content == ''
 
-    theme = (load_theme node.document)
-    math_font = node.document.attributes['math-font'] || MATHJAX_DEFAULT_FONT_FAMILY
-
-    svg_output, error = stem_to_svg(latex_content, true, math_font)
-    if svg_output.nil? || svg_output.empty?
+    if svg_output == ''
       logger.warn "Error processing stem: #{error || 'No SVG output'}"
       return super
     end
-    adjusted_svg, svg_width = adjust_svg_to_match_text(svg_output, node, theme)
-    tmp_svg = Tempfile.new(%w[stem- .svg])
-    self.class.tempfiles << tmp_svg
-    begin
-      tmp_svg.write(adjusted_svg)
-      tmp_svg.close
 
-      logger.debug "Successfully embedded stem inline #{node.text} with font #{math_font} as SVG image"
-      quoted_text = "<img src=\"#{tmp_svg.path}\" format=\"svg\" width=\"#{svg_width}\" alt=\"#{node.text}\">"
-      node.id ? %(<a id="#{node.id}">#{DummyText}</a>#{quoted_text}) : quoted_text
-    rescue => e
+    # removed svg temp file handle creation
+
+    begin
+      # removed writing of adjusted svg, and closing of handle
+
+      file_path = svg_info[:file_path]
+      svg_width = svg_info[:svg_width]
+      math_font = get_math_font(node).freeze
+
+      if error.nil?
+        logger.debug "Successfully embedded stem inline #{node.text} with font #{math_font} as SVG image"
+        quoted_text = "<img src=\"#{file_path}\" format=\"svg\" width=\"#{svg_width}\" alt=\"#{node.text}\">"
+        node.id ? %(<a id="#{node.id}">#{DummyText}</a>#{quoted_text}) : quoted_text
+      end
+    rescue StandardError => e
       logger.warn "Failed to process SVG: #{e.message}"
       super
     end
@@ -106,14 +285,12 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
   private
 
   def extract_latex_content(content, type)
-    content = content.strip.gsub("&amp;", "&").gsub("&lt;", "<").gsub("&gt;", ">")
+    content = content.strip.gsub('&amp;', '&').gsub('&lt;', '<').gsub('&gt;', '>')
     case type
     when :latexmath
-      return content
+      content
     when :asciimath
-      return AsciiMath.parse(content).to_latex
-    else
-      return nil
+      AsciiMath.parse(content).to_latex
     end
   end
 
@@ -123,10 +300,12 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
 
   def stem_to_svg(latex_content, is_inline, math_font)
     js_script = File.join(File.dirname(__FILE__), '../bin/render.js')
-    svg_output, error = nil, nil
+    svg_output = nil
+    error = nil
     format = is_inline ? 'inline-TeX' : 'TeX'
     begin
-      Open3.popen3('node', js_script, latex_content, format, POINTS_PER_EX.to_s, math_font) do |_, stdout, stderr, wait_thr|
+      Open3.popen3('node', js_script, latex_content, format, POINTS_PER_EX.to_s,
+                   math_font) do |_, stdout, stderr, wait_thr|
         svg_output = stdout.read
         error = stderr.read unless wait_thr.value.success?
       end
@@ -144,20 +323,18 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
     if node_context.is_a?(Asciidoctor::Section)
       level = node_context.level.next
       theme_key = "heading_h#{level}"
-      if node_context.sectname == 'abstract'
-        theme_key = 'abstract_title'
-      end
+      theme_key = 'abstract_title' if node_context.sectname == 'abstract'
 
       font_family = theme["#{theme_key}_font_family"] || theme['heading_font_family'] || theme['base_font_family'] || FALLBACK_FONT_FAMILY
       font_style = theme["#{theme_key}_font_style"] || theme['heading_font_style'] || theme['base_font_style'] || FALLBACK_FONT_STYLE
       font_size = theme["#{theme_key}_font_size"] || theme['heading_font_size'] || theme['base_font_size'] || FALLBACK_FONT_SIZE
       font_color = theme["#{theme_key}_font_color"] || theme['heading_font_color'] || theme['base_font_color'] || FALLBACK_FONT_COLOR
     elsif node_context
-      if node_context.parent.is_a?(Asciidoctor::Section) && node_context.parent.sectname == 'abstract'
-        theme_key = :abstract
-      else
-        theme_key = :base
-      end
+      theme_key = if node_context.parent.is_a?(Asciidoctor::Section) && node_context.parent.sectname == 'abstract'
+                    :abstract
+                  else
+                    :base
+                  end
 
       font_family = nil
       font_style = nil
@@ -183,9 +360,7 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
     end
 
     font = TTFunk::File.open(font_file)
-    unless font
-      raise "Failed opening font file: #{font_file}"
-    end
+    raise "Failed opening font file: #{font_file}" unless font
 
     descender_height = font.horizontal_header.descent.abs
     ascender_height = font.horizontal_header.ascent.abs
@@ -194,7 +369,9 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
     unless x_height
       logger.debug "'OS/2' table not found, falling back to estimating font x-height (ex) from glyph"
 
-      cmap_table = font.cmap.tables.find { |table| table.format == 4 && table.platform_id == 3 && table.encoding_id == 1 || table.encoding_id == 10 }
+      cmap_table = font.cmap.tables.find do |table|
+        table.format == 4 && table.platform_id == 3 && table.encoding_id == 1 || table.encoding_id == 10
+      end
       raise 'No suitable Unicode cmap table found' unless cmap_table
 
       glyph_id = cmap_table.code_map['x'.ord]
@@ -214,9 +391,9 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
     embedding_text_baseline_height = descender_height / units_per_em * font_size
 
     svg_doc = REXML::Document.new(svg_content)
-    svg_width = svg_doc.root.attributes['width'].to_f * POINTS_PER_EX || raise("No width found in SVG")
-    svg_height = svg_doc.root.attributes['height'].to_f * POINTS_PER_EX || raise("No height found in SVG")
-    view_box = svg_doc.root.attributes['viewBox']&.split(/\s+/)&.map(&:to_f) || raise("No viewBox found in SVG")
+    svg_width = svg_doc.root.attributes['width'].to_f * POINTS_PER_EX || raise('No width found in SVG')
+    svg_height = svg_doc.root.attributes['height'].to_f * POINTS_PER_EX || raise('No height found in SVG')
+    view_box = svg_doc.root.attributes['viewBox']&.split(/\s+/)&.map(&:to_f) || raise('No viewBox found in SVG')
     svg_inner_offset = view_box[1]
     svg_inner_height = view_box[3]
 
@@ -224,8 +401,8 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
 
     # Adjust SVG height and width so that math font matches embedding text
     scaling_factor = font_size.to_f / svg_default_font_size
-    svg_width = svg_width * scaling_factor
-    svg_height = svg_height * scaling_factor
+    svg_width *= scaling_factor
+    svg_height *= scaling_factor
 
     svg_height_difference = embedding_text_height - svg_height
     svg_relative_height_difference = embedding_text_height / svg_height
@@ -246,7 +423,7 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
       logger.debug("svg_inner_offset = #{svg_inner_offset}, svg_inner_height = #{svg_inner_height}, svg_inner_offset_new = #{svg_inner_offset_new}, svg_inner_height_new = #{svg_inner_height_new}")
       logger.debug("svg_inner_offset_diff = #{svg_inner_offset - svg_inner_offset_new}, svg_inner_offset_diff_relative = #{(svg_inner_offset - svg_inner_offset_new) / svg_inner_height}")
 
-      svg_height = svg_height * svg_inner_height_relative_difference
+      svg_height *= svg_inner_height_relative_difference
       svg_inner_height = svg_inner_height_new
       svg_inner_offset = svg_inner_offset_new - svg_inner_height_padding
     else
@@ -266,7 +443,7 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
     svg_output = adjust_svg_color(svg_doc.to_s, font_color)
 
     [svg_output, svg_width]
-  rescue => e
+  rescue StandardError => e
     logger.warn "Failed to adjust SVG baseline: #{e.full_message}"
     nil # Fallback to the original if adjustment fails
   end
@@ -274,8 +451,17 @@ class AsciidoctorPDFExtensions < (Asciidoctor::Converter.for 'pdf')
   def find_font_context(node)
     while node
       return node unless node.is_a?(Asciidoctor::Inline)
+
       node = node.parent
     end
     node
   end
 end
+
+puts("\n")
+puts('-- PATCHED with caching AsciiDoctor-PDF-MathJax extension loaded --')
+puts("\n")
+puts('To enable caching either: a) Add to your .adoc file header the attribute :' + ATTRIBUTE_CACHE_DIR + ': <Your Cache Directory>')
+puts('Or, b) Add to the AsciiDoctor-PDF command line: -a ' + ATTRIBUTE_CACHE_DIR + '=<Your Cache Directory>')
+puts('The first build of a file will take the longest because the cache is empty.  Subsequent builds will be significantly faster."')
+puts("\n")
